auth: merge "main" code into new auth-main.(cc|hh)
Merge common_startup.cc and receiver.cc into auth-main.cc, and
rename common_startup.hh into receiver.hh.
This is a very minimal merge with no cleanup. Its intention is to
make understanding the startup code path easier, by avoiding having
to look at two compilation units to determine the exact order.
