+++ pdns/auth-main.hh
